% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{BUFFER_FRAC}
\alias{BUFFER_FRAC}
\title{Default padding parameter for function get_vectors()}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
BUFFER_FRAC
}
\description{
A numeric value giving the padding offset proportion when defining 
the range for x and y coordinates
}
\keyword{internal}
