% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vectors.R
\name{.check_valid_names}
\alias{.check_valid_names}
\title{helper function to check the names of gene/cluster/sample}
\usage{
.check_valid_names(x, x_name)
}
\arguments{
\item{x}{A character vector to check naming}

\item{x_name}{A name specifying the type of x, for message purpose only}
}
\value{
A character vector of same length with valid names
}
\description{
helper function to check the names of gene/cluster/sample
}
