% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vectors.R
\name{.convert_data}
\alias{.convert_data}
\title{Convert SingleCellExperiment/SpatialExperiment/SpatialFeatureExperiment 
objects to list object for jazzPanda.}
\usage{
.convert_data(x, sample_names, test_genes)
}
\arguments{
\item{x}{a SingleCellExperiment or SpatialExperiment or 
SpatialFeatureExperiment object}

\item{sample_names}{a vector of strings giving the sample names}

\item{test_genes}{A vector of strings giving the name of the genes you want
to create gene vector.}
}
\value{
outputs a list object with the following components
\item{trans_lst }{ A list of named dataframes. Each dataframe refers to 
one sample and shows the transcript detection coordinates for each gene. 
The name matches the input sample_names}
\item{cm_lst }{A list of named dataframes containing the count matrix for
each sample. The name matches the input sample_names}
}
\description{
This function takes an object of class SingleCellExperiment, 
SpatialExperiment or SpatialFeatureExperimentreturns and returns a list
object that is expected for the \code{get_vector} functions.
}
