% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_mg_result.R
\name{.create_cor_mg_result}
\alias{.create_cor_mg_result}
\title{Create a marker gene result object for correlation approach}
\usage{
.create_cor_mg_result(obs.stat, perm.pval, perm.pval.adj)
}
\arguments{
\item{obs.stat}{A matrix containing the correlation coefficients for 
each pair of genes and cluster vectors.}

\item{perm.pval}{A matrix containing the raw permutation p-value for 
each pair of genes and cluster.}

\item{perm.pval.adj}{A matrix containing the adjusted permutation p-value
for each pair of genes and cluster.}
}
\value{
An S3 object of class 'cor_mg_result' which includes three 
matrices.
}
\description{
This function creates a structured output object named 'cor_mg_result' for 
storing the permutation results. 
The object contains three matrices:
}
