% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_mg_result.R
\name{.create_lm_mg_result}
\alias{.create_lm_mg_result}
\title{Create a marker gene result object for linear modelling approach}
\usage{
.create_lm_mg_result(top_result, full_result)
}
\arguments{
\item{top_result}{A data frame containing top results.}

\item{full_result}{A data frame containing full results.}
}
\value{
An S3 object of class 'glm_mg_result' which includes both 
results data frames.
}
\description{
This function creates a structured output object named 'glm_mg_result' for 
storing the marker gene results. 
The object contains two data frames: top results and full results.
}
