% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vectors.R
\name{get_gene_vectors_tr}
\alias{get_gene_vectors_tr}
\title{Create spatial vectors for genes from transcript coordinates}
\usage{
get_gene_vectors_tr(
  trans_lst,
  test_genes,
  bin_type,
  bin_param,
  bin_length,
  w_x,
  w_y
)
}
\arguments{
\item{trans_lst}{If specified, it is a list of named dataframes. 
Each dataframe refers to one sample and shows the transcript detection 
coordinates for each gene. Optional parameter.}

\item{test_genes}{A vector of strings giving the name of the genes you 
want to test. This will be used as column names for one of the result matrix
\code{gene_mt}.}

\item{bin_type}{A string indicating which bin shape is to be used for
vectorization. One of "square" (default), "rectangle", or "hexagon".}

\item{bin_param}{A numeric vector indicating the size of the bin. If the
\code{bin_type} is "square" or "rectangle", this will be a vector of length
two giving the numbers of rectangular quadrats in the x and y directions. If
the \code{bin_type} is "hexagonal", this will be a number giving the side
length of hexagons. Positive numbers only.}

\item{bin_length}{A positive integer giving the length of total bins}

\item{w_x}{A numeric vector of length two specifying the x coordinate
limits of enclosing box.}

\item{w_y}{A numeric vector of length two specifying the y coordinate
limits of enclosing box.}
}
\value{
a matrix contains the transcript count in each grid.
Each row refers to a grid, and each column refers to a gene.
}
\description{
This function will build gene vectors based on the transcript coordinates of
every gene
}
