% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jazzpanda-package.R
\docType{data}
\name{rep1_neg}
\alias{rep1_neg}
\title{Rep1 negative control genes within the selected region.}
\format{
A SpatialExperiment object. 
The molecules assay slot is a BumpyDataFrameMatrix obejct. 
Can retrieve DataFrame version by calling
`BumpyMatrix::unsplitAsDataFrame(molecules(rep1_neg))`. 
The molecules slot contains: 
\describe{
    \item{x}{x coordinates}
    \item{y}{y coordiantes}
    \item{feature_name}{negative control probe name}
    \item{category}{negative control category}
}
}
\source{
<https://cf.10xgenomics.com/samples/xenium/1.0.1/
Xenium_FFPE_Human_Breast_Cancer_Rep1/Xenium_FFPE_Human_Breast_
Cancer_Rep1_outs.zip>
}
\usage{
data(rep1_neg)
}
\value{
SpatialExperiment
}
\description{
A SpatialExperiment object containing the coordinates for every negative 
control detection for rep1_sub
}
\keyword{datasets}
