% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jazzpanda-package.R
\docType{data}
\name{rep2_sub}
\alias{rep2_sub}
\title{A small section of Xenium human breast cancer rep2.}
\format{
A data frame with 90375 rows and 3 variables:
\describe{
    \item{x}{x coordinates}
    \item{y}{y coordiantes}
    \item{feature_name}{transcript name}
}
}
\source{
<https://cf.10xgenomics.com/samples/xenium/1.0.1/
Xenium_FFPE_Human_Breast_Cancer_Rep2/
Xenium_FFPE_Human_Breast_Cancer_Rep2_outs.zip>
}
\usage{
data(rep2_sub)
}
\value{
List
}
\description{
A small section of Xenium human breast cancer rep2.
}
\keyword{datasets}
